Update RTGQueueSystem.py
Started to implement a system where an RTG is able to go on tour in the wrong order and then be be placed back in the correct   position of the roster. However, it does not work due to the fact that for the RTG to be placed back on the roster in their original position the RTG ahead of them must not be on tour. Though, when thinking more about the issue it makes more sense for the RTG to always be put to the back of the queue when returning from tour